requirements.txt
streamlit==1.38.0
redis==5.0.4
rq==1.16.2
insightface==0.7.3
onnxruntime==1.18.0
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
scikit-learn==1.5.1
python-dotenv==1.0.1
openpyxl==3.1.5
pillow==10.4.0
pillow-heif==0.22.0
pillow-avif-plugin==1.5.2
pyyaml==6.0.2
scipy==1.13.1
lap==0.5.12
plotly==5.24.1
